Updates for move to IBM namespace
This changes the project to be a small wrapper
around the redstone library which is where most of the
development focus for the Python KeyProtect client has
moved.

This will remain a small wrapper to provide users an easy
way to find a Python lib that can provide the simple interface
that they are most likely looking for, but not have multiple copies
of code sitting around creating a code maintainence nightmare.